+++ flow-typed/class-name-strategy.typedef.js
@@ -1,0 +1,3 @@
+/*::
+declare type ClassNameStrategy = (componentData: any, theme: any, styles: any, isMobile: boolean) => string;
+*/