+++ packages/common/web/src/Utils/classNameStrategies.js
@@ -1,7 +1,9 @@
+// @flow
+
 import classNames from 'classnames';
 import { camelCase, upperFirst } from 'lodash';
 
-export const alignmentClassName = (componentData, theme, styles) => {
+export const alignmentClassName /*: ClassNameStrategy*/ = (componentData, theme, styles) => {
   const { alignment } = componentData.config || {};
   if (!alignment) {
     return '';
@@ -10,7 +12,7 @@ export const alignmentClassName = (componentData, theme, styles) => {
   return classNames(styles[key], theme[key]);
 };
 
-export const sizeClassName = (componentData, theme, styles) => {
+export const sizeClassName /*: ClassNameStrategy*/ = (componentData, theme, styles) => {
   const { size } = componentData.config || {};
   if (!size) {
     return '';
@@ -19,7 +21,7 @@ export const sizeClassName = (componentData, theme, styles) => {
   return classNames(styles[key], theme[key]);
 };
 
-export const textWrapClassName = (componentData, theme, styles) => {
+export const textWrapClassName /*: ClassNameStrategy*/ = (componentData, theme, styles) => {
   const { textWrap } = componentData.config || {};
   if (!textWrap) {
     return '';