+++ packages/editor/web/src/RichContentEditor/Toolbars/StaticToolbar/createFooterToolbar.jsx
@@ -6,6 +6,7 @@ const getFooterTheme = theme => {
   const { toolbarStyles: toolbarTheme, buttonStyles, separatorStyles: separatorTheme, ...rest } =
     theme || {};
 
+  /* eslint-disable @typescript-eslint/camelcase, camelcase*/
   return {
     toolbarStyles: {
       toolbar: classNames(toolbarStyles.footerToolbar, toolbarTheme && toolbarTheme.footerToolbar),
@@ -33,12 +34,10 @@ const getFooterTheme = theme => {
         toolbarStyles.footerToolbar_responsiveArrowEnd,
         toolbarTheme && toolbarTheme.footerToolbar_responsiveArrowEnd
       ),
-      //eslint-disable-next-line camelcase
       responsiveArrowStart_icon: classNames(
         toolbarStyles.footerToolbar_responsiveArrowStart_icon,
         toolbarTheme && toolbarTheme.footerToolbar_responsiveArrowStart_icon
       ),
-      //eslint-disable-next-line camelcase
       responsiveArrowEnd_icon: classNames(
         toolbarStyles.footerToolbar_responsiveArrowEnd_icon,
         toolbarTheme && toolbarTheme.footerToolbar_responsiveArrowEnd_icon