+++ packages/editor/web/src/RichContentEditor/Toolbars/default-toolbar-settings.js
@@ -1,3 +1,4 @@
+//@flow
 import { TOOLBARS, DISPLAY_MODE } from 'wix-rich-content-editor-common';
 import { createSideToolbar } from './SideToolbar';
 import { createMobileToolbar, createFooterToolbar, createStaticTextToolbar } from './StaticToolbar';
@@ -48,7 +49,11 @@ const defaultTextPluginButtons = {
   },
 };
 
-export const getDefaultToolbarSettings = ({ pluginButtons, textButtons, pluginTextButtons }) => {
+export const getDefaultToolbarSettings /*: GetToolbarSettings*/ = ({
+  pluginButtons,
+  textButtons,
+  pluginTextButtons,
+}) => {
   return [
     {
       name: TOOLBARS.SIDE,
@@ -118,7 +123,7 @@ export const getDefaultToolbarSettings = ({ pluginButtons, textButtons, pluginTe
         };
       },
       getTextPluginButtons: () => ({
-        desktop: [],
+        desktop: {},
         mobile: {
           ios: pluginTextButtons.mobile,
           android: pluginTextButtons.mobile,