+++ packages/plugin-image/web/src/toolbar/insert-buttons.jsx
@@ -1,8 +1,9 @@
+// @flow
 import { TOOLBARS } from 'wix-rich-content-editor-common';
 import { DEFAULTS } from '../consts';
 import { InsertPluginIcon } from '../icons';
 
-export default ({ helpers, t, settings }) => {
+const createInsertButtons /*: CreateInsertButtons */ = ({ helpers, t, settings }) => {
   const icon = settings?.toolbar?.icons?.InsertPluginButtonIcon || InsertPluginIcon;
   return [
     {
@@ -18,3 +19,5 @@ export default ({ helpers, t, settings }) => {
     },
   ];
 };
+
+export default createInsertButtons;