+++ packages/plugin-link/web/src/toolbar/inline-buttons.jsx
@@ -1,15 +1,17 @@
+// @flow
 import React from 'react';
 import { BUTTONS, updateLinkAtCurrentSelection, EditIcon } from 'wix-rich-content-editor-common';
 import TextLinkButton from './TextLinkButton';
 import RemoveLinkButton from './RemoveLinkButton';
 import UrlLinkButton from './UrlLinkButton';
 
-export default (config, closeInlinePluginToolbar) => {
+const createInlineButtons /*: CreateInlineButtons */ = config => {
   return [
     {
       keyName: 'url',
       component: props => <UrlLinkButton {...config} {...props} />,
       mobile: true,
+      type: BUTTONS.CUSTOM,
     },
     { keyName: 'separator1', type: BUTTONS.SEPARATOR, mobile: true },
     {
@@ -20,27 +22,26 @@ export default (config, closeInlinePluginToolbar) => {
             insertLinkFn={updateLinkAtCurrentSelection}
             icon={EditIcon}
             tooltipText={config.t('LinkTo_Edit_Tooltip')}
-            closeInlinePluginToolbar={closeInlinePluginToolbar}
             {...config}
             {...props}
           />
         </div>
       ),
       mobile: true,
+      type: BUTTONS.CUSTOM,
     },
     { keyName: 'separator2', type: BUTTONS.SEPARATOR, mobile: true },
     {
       keyName: 'remove',
       component: props => (
         <div style={{ margin: '0 -6px 0 -6px' }}>
-          <RemoveLinkButton
-            closeInlinePluginToolbar={closeInlinePluginToolbar}
-            {...config}
-            {...props}
-          />
+          <RemoveLinkButton {...config} {...props} />
         </div>
       ),
       mobile: true,
+      type: BUTTONS.CUSTOM,
     },
   ];
 };
+
+export default createInlineButtons;