+++ packages/plugin-link/web/src/typeMapper.js
@@ -1,7 +1,9 @@
-import { EXTERNAL_LINK_TYPE, LINK_TYPE } from './types';
+// @flow
+//import { EXTERNAL_LINK_TYPE, LINK_TYPE } from './types';
 import LinkViewer from './LinkViewer';
 
-export const typeMapper = () => ({
-  [EXTERNAL_LINK_TYPE]: { component: LinkViewer, elementType: 'inline' },
-  [LINK_TYPE]: { component: LinkViewer, elementType: 'inline' },
+// if [LINK_TYPE] key is used, flow won't typecheck the value. See and upvote: https://github.com/facebook/flow/issues/4649
+export const typeMapper /*: PluginTypeMapper*/ = () => ({
+  'wix-draft-plugin-external-link': { component: LinkViewer, elementType: 'inline' },
+  LINK: { component: LinkViewer, elementType: 'inline' },
 });